* 64-bit code cleanups
git-svn-id: http://gimp-normalmap.googlecode.com/svn/trunk@46 e7907806-2427-11df-9d2b-970631db4c00

+++ normalmap.c
@@ -1168,9 +1168,9 @@ static gint idle_callback(gpointer data)
 
 static void filter_type_selected(GtkWidget *widget, gpointer data)
 {
-   if(nmapvals.filter != (gint)(long)data)
+   if(nmapvals.filter != (gint)((size_t)data))
    {
-      nmapvals.filter = (gint)(long)data;
+      nmapvals.filter = (gint)((size_t)data);
       update_preview = 1;
    }
 }
@@ -1191,9 +1191,9 @@ static void height_source_selected(GtkWidget *widget, gpointer data)
 {
    GtkWidget *opt;
    
-   if(nmapvals.height_source == (gint)(long)data) return;
+   if(nmapvals.height_source == (gint)((size_t)data)) return;
    
-   nmapvals.height_source = (gint)(long)data;
+   nmapvals.height_source = (gint)((size_t)data);
    
    opt = g_object_get_data(G_OBJECT(widget), "conversion_opt");
    if(!nmapvals.height_source)
@@ -1212,9 +1212,9 @@ static void height_source_selected(GtkWidget *widget, gpointer data)
 
 static void alpha_result_selected(GtkWidget *widget, gpointer data)
 {
-   if(nmapvals.alpha != (gint)(long)data)
+   if(nmapvals.alpha != (gint)((size_t)data))
    {
-      nmapvals.alpha = (gint)(long)data;
+      nmapvals.alpha = (gint)((size_t)data);
       update_preview = 1;
    }
 }
@@ -1223,9 +1223,9 @@ static void conversion_selected(GtkWidget *widget, gpointer data)
 {
    GtkWidget *contrast_spin;
    
-   if(nmapvals.conversion != (gint)(long)data)
+   if(nmapvals.conversion != (gint)((size_t)data))
    {
-      nmapvals.conversion = (gint)(long)data;
+      nmapvals.conversion = (gint)((size_t)data);
       contrast_spin = g_object_get_data(G_OBJECT(widget), "contrast_spin");
       gtk_widget_set_sensitive(contrast_spin, nmapvals.conversion == CONVERT_HEIGHTMAP);
       update_preview = 1;
@@ -1248,9 +1248,9 @@ static void dudv_selected(GtkWidget *widget, gpointer data)
    GimpDrawable *drawable;
    GtkWidget *opt;
    
-   if(nmapvals.dudv == (gint)(long)data) return;
+   if(nmapvals.dudv == (gint)((size_t)data)) return;
    
-   nmapvals.dudv = (gint)(long)data;
+   nmapvals.dudv = (gint)((size_t)data);
    
    drawable = g_object_get_data(G_OBJECT(widget), "drawable");
    opt = g_object_get_data(G_OBJECT(widget), "alpha_opt");
